ENH: The axis component of the parameters are now normalized before creating the quaternion.
In the Jacobian, derivatives with respect to the axis components are
multiplied by the normalization factor to convert it back into
non-normalized space.

To make the Jacobian calculations more efficient, some results
are cached when SetParameters is evoked.

The derivatives of the axis components is also multiplied by
a TranslationScale to aid optimization of the axis components.