ENH: updated this NrrdIO to match latest Teem (version 1.9, to be released mid-July).  The significant upgrade here is the ability to read the "measurement frame" field in NRRD files (with magic "NRRD0005"), corresponding to a current NAMIC standard for describing diffusion-weighted MRI measurements.
Also, in order to decreasing importance

- begin the task of mapping between nrrd orientation and ITK
orientation.  The mapping from nrrd to ITK works, the other direction
will be finished soon.

- the measurement frame will be stored in the MetaData directionary;
this will be finished soon.

- non-scalar pixel types are still not handled

- removed all things related to "BigBitField", platform-specific
detail that is no longer needed for compilation

- added support for basic operations: nrrdSlice, nrrdCrop,
nrrdAxesPermute, and nrrdShuffle, which may come in handy for mapping
memory layouts between nrrd and ITK as part of input/output