COMP:  Compiler warning suppression.  The copy constructor needs to be public.
"/buildtemp/Insight/Testing/Code/Algorithms/itkCommandIterationUpdate.h", line 116: Warning (Anachronism): Formal argument event of type const itk::EventObject& in call to itk::Object::AddObserver(const itk::EventObject&, itk::Command*) has an inaccessible copy constructor.

===========
--  clip from http://forum.java.sun.com/thread.jspa?threadID=5094344 --
In patches and new compiler releases we try to improve the quality of the compiler. Some enhancements consist of finding errors that earlier compiler versions missed -- often a due to a bug report from a customer.

An "Anachronism" warning refers to code that was allowed (correctly or incorrectly) by compilers implementing old-style C++, before the 1998 C++ Standard.

In cases where we newly detect an error that is not necessarily dangerous, we usually emit a warning instead of an error, so that working code will continue to compile.

But the warning is telling you that the code is not valid, and is likely to be rejected by other compilers, or by later Sun compilers.

In this case, the C++ language rule is that the copy constructor of A must be accessible when passing a temporary A object, even to a reference parameter. The reason is that implementations are allowed to make extra copies of the o
bject, requiring the copy constructor. If objects of type A really should not be copied, your program could fail. The Sun compiler does not actually make an extra copy, so probably no harm is done by allowing the code.

Whenever you get an Anachronism warning, you should fix your code so that no warning is issued. In this case, the fix is simple: pass a real object instead of a temp: [code]
A tmp;
test(tmp); [/code]
--- clip
refer to the section number in ANSI C++ standard?
--- clip
the last sentence in 8.5.3, p5, bullet 2: "...The constructor that would be used to make the copy shall be callable whether or not the copy is actually done."
===========