Removed Brains2Mask reader/writer
The Brains2Mask file format has very limited functionality
outside the legacy brains2 program.  This file format
does not have sufficient information for supporting
orientation of masks, and even the Iowa contributors
have given up on using the ITK Brains2Mask as a file
format from ITK.  This is a legacy format that was
never included in the default file reader lists.