Check for local Git hooks on first commit
Install a small .git/hooks/pre-commit script in the source tree at CMake
configuration time.  When the user first attempts to create a commit
this bootstrapping hook will trigger and print instructions to install
the real hooks.  This approach allows us to require local hooks for
commit but not for those who just want to build.