ITK_HAS_INT_64 has been removed. Assumed to have a 64-bit integer.
Mantis Bug 0011081: ITK_HAS_INT_64 should be removed.

When this flag was added only bcc did not have a functioning 63-bit
integer. With the Borland compiler being remove this flag should also
be removed.

This flag was defined in itkIntTypes.h and needed for rounding
methods. These file now assume that a int64_t is defined.