Updated openjepg to v2. Disabled j2k in gdcm.
The third party library openjpeg has been updated to the most recent
version from their SVN repository. This version is incompatible with
our current version of GDCM, so temporarily we have disabled the
use of j2k from gdcm. This will be fixed as soon as GDCM 2.0 is
introduced in ITKv4.