Moved JPEG2000Image from IJ to Review.
The contribution of the JPEG2000ImageIO class from the
Insight Journal:

  "Support for Streaming the JPEG2000 File Format"
  by Mosaliganti K., Ibanez L., Megason S.
  Harvard University, Kitware

  http://hdl.handle.net/10380/3187
  http://www.insight-journal.org/browse/publication/741

has now been moved to the Review directory.

The test code is included, but no tests have been defined
yet because some of the needed input images are quite large.
That said, being able to manage large images should be
part of the tests. We should discuss this during next tcon.