IdentityTransform now reports the correct number of parameters.
The IdentityTransform was being initialized as though it had 1 parameter.
It is now initialized to have 0 parameters. The test has been updated
to reflect this change so that it does not check specific elements values
of the Jacobian, checks that the number of columns of the Jacobian is 0,
and checks that the number of parameters is reported as 0. Also, extra
spaces at the end of lines in the test were removed.

Change-Id: I9ccf368ccc34401f1012d68663e48d81f645e9c2