Statistics Tests were disabled.
During the transition of the Statistics Framework from the Review
directory, the Statistics tests were disabled (temporarily).
Here we re-enable these tests. In the process we moved the file
itkTestingMacros.h from the Testing/Code/Review directory to the
Code/Common directory, since it is used by the Statistics tests.

Change-Id: Ic2d0dbd7bb33f1a96e09e78e9196cca0c77f853f