Statistics Tests were disabled.
During the transition of the Statistics Framework from the Review
directory, the Statistics tests were disabled (temporarily).
Here we re-enable these tests. In the process we moved the file
itkTestingMacros.h from the Testing/Code/Review directory to the
Code/Common directory, since it is used by the Statistics tests.

Change-Id: Id09bdb42e56e27758d8ff31e42f008ec3efbe33a