Add empirical style specifications to uncrustify conf.
Some types of options in the uncrustify style configuration file are set to
values empirically consistent with ITK coding style that took place
over the first 10 years of development.

To determine the boolean options, the configuration file is set to each value, and
the number of line changes required to enforce the value are counted.  The
value (false or true) with the lowest number of line changes is used.

A similar process is applied with the 'add/remove' and 'lead/trail' options.
 However, we also calculate a 'percent consensus' for these options by
examining the number of line changes required when the value is set to
'ignore'.  If we take 'a' to be the number and line changes for the 'add'
value, 'r' for 'remove', and 'i' for 'ignore', the consensus percent is

  consensus_percent = 100.0 * ( 1.0 - min( a-i, r-i ) / ( a-i + r-i ))

For these options we require the consensus percent to be above a threshold,
or the value is set to 'ignore'.

Tradition suggests a 2/3's majority, 66.7% consensus, should be used.
 However, if we look at a histogram of the consensus percent values
(attached), many more options can be specified by lowering the threshold.
 So, 57% is used.

My limited experience so far has been that an uncrustify will make a file
pass KWStyle test without problem apart from higher level issues like header
guards or missing namespaces.  Again, the purpose is speed up development
while retaining decent looking code.  As Hans and Brad remark, allowing
changes that are only style-related will make version history messy.  When
using the merge tool, the default merge file does not have any uncrustify
changes.  Only changes related to the commit should be brought in.  Commits
should be stylish content as opposed to style only.  Think The Artist
Formerly Known as Prince as opposed to Lady Gaga.  Or something like that
;-).

Note that the configuration file was also updated to uncrustify version 0.56 by
using the uncrustify option --update-config-with-doc.

These empirically determined values are found in
'uncrustify_itk_aggressive.cfg', while uncrustify_itk.cfg contains enough
specified values so the result will pass the KWStyle test.  uncrustify_itk.cfg
was set according to the specifications found in Documentation/Style.pdf.

Change-Id: Ib390bfdc4e570d16d8ac9c6b11da40dbf98903ce