Move CMake commands to match new conventions
The next step in removing end-of-lines is to run through all the
CMakeLists.txt files and clean them up.

This also changes all the key words to lower
case. The primary reason for changing key words is that all documentation for
CMakeLists.txt now shows the key words as lower case. Even the printed
“Mastering CMake v5” uses lower case. I’ve come across this several
times in my class where students stumble over the fact that the
documentation about Cmake does not match the implementation in ITK.

I have a script (a wrapper around a vim macro) written that can make
these substituions reliably (Tested on the BRAINS tree, the Slicer3
tree, and the ITK tree without causing any errors).  more complete
documentaiton can be found attached to

http://public.kitware.com/Bug/view.php?id=11175