Removed not needed function and includes.
Removed extra included headers.

Remove GenerateInputRequestedRegion since the default behavior of the
superclass performs the correct operation and this methods is not
needed. By default, the output region is copied to the input
region. Since the output region is set to the largest possible region
so will the input. This method did not need to be overridden.

Change-Id: Ie68d47486f58b715d50e9f84b15143bbeb74f4b7