enhanced perimeter estimation
The perimeter estimation is not using the raw image anymore. Everything
is computed from the LabelObject RLE. The memory usage is decreased
and the code is multithreaded. The accuracy of the algorithm has been
improved by using more directions to estimate the perimeter in 2D and
3D. The perimeter now include the contour of the object touching the border
of the image and is now estimated by default.

Change-Id: I1f247f67090ed0e8e50209bc6021e2563c865e08