NumericTraits - Added MeasurementVectorType to all NumericTraits
For the generic NumericTraits<T>, MeasurementVectorType is defined as T[1]. For
the container types, it is defined as Self. GetLength and SetLength are
defined for T[1] so that GetLength will always return 1 and SetLength will
throw an exception unless the length to be set is 1 in which case it will reset
the single array element to the zero value of the element type.

In adition, the MeasurementVectorTraits test was revived and a test of the
GetLength and SetLength properties for the new MeasurementVectorTypes was added

NumericTraits - Use forward declaration to get around include loop

Using Brad Lowekamp's patch to circumvent the include loop created by
including FixedArray in NumericTraits.

Change-Id: Ia061d35c8d4a2747fd3eb389ecd0166372903a7c