fix the pipeline behavior of InPlaceLabelMapFilter
After an Update(), when run in place, the filter was run again
by a second Update(). This behavior was triggering unpredictable
behaviors in ObjectByObjectLabelMapFilter, by updating the content
of the label object container and thus invalidating the iterator
used in that filter.
The method removed was copied from InPlaceImageFilter but doesn't
seem to make much sense in the LabelMap case.

Change-Id: I1f1617d508c023d4b37d42b211e8118ccc2b5e6f