Fixed files that broke during regex replacement.
The regular expression of the python script used to replace the
copyright notices, got confused in some particular cases in which
C++ comments (similar to the ones of the copyright header) were
also located at the middle and at the end of the files. This
commit manually corrects those cases.

Change-Id: I1b205cc9d7f8a7b4891ea82479bfa80afb4d5d6e