Avoiding problem with relative path.
Then running the Nightly build of the Testing directory against
and installed ITK. This file doesn't pull the CTestConfig.cmake
file in the top directory of ITK. It looks like the ITKTesting_SOURCE_DIR
cmake variable is not defined yet at the time that this file is
viewed by ctest.

We avoid the problem by simply copying the content of CTestConfig.cmake.
It is not an ideal solution, since now the two files can go out of sync.
Exploring the perfect solution may require assistance from CMake Gurus.

Change-Id: Ia59399e6ded22e6cc4cd2ec015488de34ca37f87