BUG: Remove NumericTraits from Array that require a Self(X).
Since itk::Array is runtime length determined object, it does not have a
contructor that takes a single default value for the entry contents.  As such,
it should not have NumericTraits defined for it such as
NumericTraits<Array<T>>::OneValue().  The entries have been removed and replaced
with the similar implementation in itkNumericTraitsVariableLengthVectorPixel.h.

Change-Id: I3eddf3d575fcc392a22a99825852d482d4faddbe