BUG: Fixed RecursiveGaussian filter to normalize across scale-space
For the 3D case the laplacian was incorrectly normalized with sigma^3
not sigma^2. This error occours in the base class
itkRecursiveGaussianImageFilter, and effects all recursive gaussian
filters with normalize across scale option.

From my recent readings there is a "natural coordinates" or
"dimensionless spatial units" which are frequently used for this
analysis. The conclusion I reached from this analysis is that the
correct scaling factor should be sigma^n where n is the order of the
differentiating for the separated convolutions. ( Not the current
constant sigma ).

Additionally the scale space normalization was also scaled by the
pixel spacing which is also incorrect.

ENH: Added test for scale-scale normalization gaussian

Compute gaussians of a sin function to compare results to theoretical
values and verify correct behavior.

DOC: Updated documentation for NormalizeAcrossScale methods

BUG: updated test to work with corrects scale-space normalization

The example ResampleVolumesToBeIsotropic does not perform scale-space
analisys, so it is no longer set. The baseline image needs to be
updated.

HessianRecursiveGaussianFilterScaleSpaceTest's expected values needed
to be updated for corrected normalization with respect to pixel
spacing.

RecursiveGaussianImageFiltersTest was updated to check that no
normalization occours with Zero order convolutions.

ENH: added energy conservation test to recursive gaussian test

The energy of the input image of should be the same and the energy of
the output image ( ignoring the boundary condition ). The test added
verifies the energy of an impulse is approximatly preserved after the
filter.

Change-Id: I2251f658aea186a5919535129e077401904cd77e