BUG: 2899 Fixed default output image type.
The output image should be defined in terms of the input image traits.
Note that the new default is not perfect either, because it assumes
the existance of the "itk::Image" type, instead of allowing for a fully
generic class that just provides the API of the itk::Image.

Also included here some Style changes that were required by KWStyle
at commit time.

Change-Id: I43c4c6a6d76b5b4bd51c63def273da9f88e5847c