ENH:  Find correct version of zlib (11493)
When installed versions of ITK are used,
there needs to be a way of identifying
the correct header to use for name
mangled versions of the zlib.

STYLE:  Final style changes for nifti merge

After much discussion regarding the importance
and need for consistency in const correctness
it was decided that the version of nifti inside
of ITK will have a const-consistent set of
functions, and nifti ouside of ITK will deliver
compile time warnings where the const is not
used every where that it is needed.

This patch also removes end-of-of line spaces.

Change-Id: I65f65e545a9621a6bc11baa96a4d76c3cc565938