BUG: 8893 Making internal Mean types to be real
The internal types used for real computation and accumulation
were not using the proper numeric traits.

Now that the mean is returned in a MeasurementVectorRealType instead
of the MeasurementVectorType of the sample type, the tests have been
modified accordingly.

This may still benefit from replacing the simple summation with
the Kahan summation as Brad suggested in the bug report.

ENH: Addressed comments from Arnaud's review.

Change-Id: I96b870dd31b34766d159c3d4a7086793ecbfec71