ENH: don't use the kernel date and its build date in the wisdom name
GetOSRelease() is actually returning the kernel version, as returned
by uname -r on linux. This is not the right behavior, as a security
update of the kernel is updating the kernel version. A simple security
update shouldn't lead to regenerating all the wisdoms. On most linux
systems, the lsb_release command would produce a more useful
description:

  [glehmann@marvin ITK (fftw)]$ lsb_release -a
  LSB Version:lsb-4.0-amd64:lsb-4.0-noarch:*
  Distributor ID:MandrivaLinux
  Description:Mandriva Linux 2010.1
  Release:2010.1
  Codename:Henry_Farman

The same is true for GetOSVersion() which returns the kernel build
date and thus is also modified by a kernel security update.

An alternative may be to fix kwsys.

Change-Id: If874c1bfc663694b4f7fbe2fc69aaea6fc3612de