ENH: Made StreamingImageIO methods depended on CanStream methods
Some ImageIOs may not be able to stream read and write all available
options. This patch let the CanStreamWrite and CanStreamRead methods
change base of the configuration of the ImageIO. Because of the order
the ImageFileReader and ImageFileWriter execute, the ImageIO will have
their configurations ( pixel type etc ) before the the region
management methods of the ImageIOBases are called.

Change-Id: I264cadb3e6e50a69919a553f0d41aa9f54560a07