BUG: BIGENDIAN config in TIFF and openjpeg
The CMAKE_WORDS_BIGENDIAN symbol was not used properly in the
configuration of the .h files. In openjpeg the file needed to
be converted to a .h.in and processed in order to generate a
file in the binary directory.  We suspect that this lack of
configurarion was the source of the failing test in big endian
architectures.

Addressed review comments:
When building in APPLE, we must take into account not just the
build system but the target system. We have now copied the logic
that is used in itkConfig.h.in and factorized some of the symbols.

Change-Id: I428825443e169ff0ebe673ac28feca8f59063a2a