BUG: Fixed output path to avoid infinite coping.
ITK_remaining dir was set to be located in the dir where the scripts are running. When the scripts
are running from the ITK/Modularization dir, the ITK_remaining dir will become part of the monolithic
ITK, which will become a problem when calling the recrusive copying (shuntil.copytree).
The problem is solved by moving the ITK_remaining dir to  the moduarlITK output dir. Output logs are
also moved to the modularITK dir.

Change-Id: Ifa19df8a78fb607ca1b518123105524b2d2114dd