BUG: Call UnRegister() after CreateInstance().
The CreateInstance() method has an unnecessary call to Register() that results
in a incorrectly high reference count for the objects created in this method.
As a consequence, memory leaks occurr in the classes that used this call.
Notably the SpatialObject duplicator and the TransformFileReader.

This extra count must be removed in order to prevent memory leaks. The option
of removing the call to Register() that is made inside of CreateInstance() will
produce a non-backward compatible behavior.

In order to solve the memory leaks, we now call UnRegister() in the newly
created objects thata are returned by CreateInstance().

Change-Id: I08862b22bccddbf3ab78ae3fa80d26d9d0d02500