BUG: Fixing memory leaks.
Rewriting the itkSimplexMeshAdaptTopologyFilter to respect the constness of the
input mesh. The previous version was modifying the input mesh as part of the
process of computing the output mesh.

ENH: Manage the case when nset is null during the copy of data structures.

BUG: Computation should operate in the output mesh. Now that we have copied the
input mesh into the output mesh, the original input mesh does not have to be
involved in the computation.

BUG: Fix memory leak in ReplaceFace() method.

BUG: Improving ReplaceFace() method with what we learned from the
itkMeshTest.cxx.

STYLE: Fixing typo in documentation.

BUG: Identified the last leak.

When deleting a cell from the mesh, two things must be done: the cell must be
removed from the cell container, and memor of the cell itself must be released
by calling "delete".

Change-Id: I0ad81eed663b50724918fdf2f8391270ce996563