ENH: Now downloading BrainWeb dataset from MIDAS.
By enabling the new CMake flag: ITK_USE_BRAINWEB_DATA, the user will schedule
the download of BrainWeb datasets from MIDAS and will enable the tests (in
Examples/Segmentation and Examples/Registration) that use those images as
input. This is intended for heavier than normal testing, particularly to
exercise ITK with realistic datasets.

Presumably this will be enabled in selected machines, or in Weekly builds
(instead of Nightly builds).

Change-Id: Ib02caa05792d71c567c73a2828d982786fdae030