ENH: Disseminated use of ITK_USE_MODULAR_BUILD.
Introduce ITK_USE_MODULAR_BUILD flag.

This flag will allow us to enable/disable code that must be different
between the monolithic build of ITK and the modularized build of ITK.
For example, by removing code that must be separated due to dependencies
of the modules in modularized ITK.

This flag is used in the codes where source code must be enabled or
disabled depending on whether the code is being built in a modular way.

Removed DICOMIO2, DicomImageIO and DICOMSeriesFileNames from modularized ITK.

Change-Id: Id6286215768b271222009d1c6fd492c1c5fb1d77