BUG: Copy of sharpenedImage was missing SetRegions().
The sharpenedImage is duplicating the grid of the input image.
When replacing Graft() with CopyInformation() we missed to add
SetRegions(). The Graft() method shouldn't be used here because
it is intended to reuse the pixel container of the grafted image.
(In this case the grafted image was the input image, which breaks
the rule that a filter shouldn't modify its input).

Change-Id: I877e717ab5e3c1ef1791d0e64a897e3aeb20921c