COMP: long double doesn't initialize all its bytes.
The long double type (in the platform where we are running valgrind),
and in Ubuntu 10.04 with gcc 4.4.3, does not initialize all its bytes
when assigning a value. It reports its sizeof() to be 16, but only
initialize 11 of those bytes. In order to get around an uninitialized
report by Valgrind, here we declare an union between the long double
and an array of chars of the same size, in order to use the char array
to initialize the same memory where the long double is allocated.

Change-Id: Ic0ea8cd2908402a0298cc7da9e070198ca684565