BUG: Mixing up # of elements vs # of bytes
The method MetaImage::M_ReadElementsROI(), takes the parameter
_totalDataQuantity which should be in the unit of the number of
pixels, not the number of bytes. Changed calls to pass to the
number of pixels not number of bytes.

ENH: Test comparison can be restored now.

The comparison with baseline was commented out in the CMakeLists.txt
file because of the incorrect output. After the fixes in metaImage.cxx
this comparison against the baseline can now be restored in the
CMakeLists.txt file.

Change-Id: I20d88eb5b1419eb0befe838f163a9ddc450757e0