ENH: modularization scripts modifications.
1. BUG: The modularITKSupport repository was only downloaded when the output
modular ITK path exsit. The assumption was wrong since users should be
able to create new directories for the output. The bug is fixed by moving
"git clone" commands outside the "if existing"condition.
2. ENH: Imporved comments in customRun.sh.
3. ENH: The tests in the Manifest file only contained covers
the test for the codes in 'src'. By adding "include" dir in the
testFinder.py, tests for the template txx files are collected.

Change-Id: Ibe2d808808b7cc310fc23f2699ef1787847f7080