COMP: ObjectFactoryBase memory leak.
LoadLibrariesInPath was not closing the library after a failing to
find a symbol in a library, causing a memory leak.

In the past, some platforms crashed if DynamicLoader::CloseLibrary was
called. This patch is intended to see if any current ITK4 platforms
crash. If there are some platforms that still crash we may be able to
ifdef the call to DynamicLoader::CloseLibrary.

Change-Id: I4905f96df8c739219869e2b8a24f5e08e19606b9