COMP: Fix warning about possible loss of data.
The type VectorContainerSizeType was replaced with ElementIdentifier.

VectorContainerSizeType was defined from the size_type of the internal
STL vector, but broke the consistency of the API with all other methods
that used ElementIdentifier as their return type. This was also making
the API of the VectorContainer::Size() method look different from the
API of the MapContainer::Size() method.

ENH: Fixed warning about comparison between signed and unsigned in
the itkPriorityQueueTest.cxx.

Change-Id: I36c63a3873855a58fd56d33d298f3f2e3670a619