BUG: 11699 IsInside and ValueAt methods give inconsistent results.
The following description and fix were provided by Andras Lasso.

ImageMaskSpatialObject::IsInside (which determines if the point is
inside or not the masked region) uses index[i] = static_cast<int>(
p[i] ) to convert from continuous index to integer index.

This is inconsistent with ImageSpatialObject::ValueAt (which returns
the pixel value at a specified position) that uses
m_Interpolator->EvaluateAtContinuousIndex(index).

The effect is that when using a mask for registration many points are
incorrectly rejected at the mask boundary during the sampling step (in
ImageToImageMetric<TFixedImage,TMovingImage>::SampleFixedImageRegion). When
the mask is thin (such as when the ROI is a single slice) this often
leads to registration failure (e.g., "Joint PDF summed to zero").

Fix: Change ImageMaskSpatialObject::IsInside to convert
continuous indices to integer indices the same way (using the
interpolator) as it is done in ImageSpatialObject::ValueAt.

Change-Id: Ie6f464c91ba312c83317bb427fd6a19eee897bb4