COMP: Incorrect SSE2 configuration.
The logic used for discovering the availability of SSE2 rounding functions was
flawed. We have now introduced an ITK native try compile for 32 and 64 bits.

Before, the problem was that ITK configuration was using a cache value from VXL
before the add_subdirectory(vxl) command. This caused the -msse2 flag to only
be added after a second run of cmake in the binary tree, and therefore this
could cause a complete rebuild.

Also added specific try compile test for checking whether the -msse2 flag is
needed at all or not by the compiler.

                               __SSE__ 1
                               __SSE2__ 1
                               __SSE2_MATH__ 1
                               __SSE_MATH__ 1

defined by default.

The flag -msse2 is, however, no longer added as a required flag.

Additional test for just "emmintrin.h" is used independetly
conditionally include in itkMathDetail.h.

Change-Id: I07e8dba64b8b06b1064a350e5a13e60832fb1ab9