ENH: enabled io pasting for image not connected to pipeline
For io pasting the largest possible region of the image, must match
that in the file. However, if an input image has no source filter,
then the largest possible region would be changed to the buffered
region in the UpdateOutputInformation method. This prevented using the
pasting with out the input to the write fully connected a pipeline.

The fix is to special case the pipeline execution in the
ImageFileWriter for this special condition.

ENH: test to perform io pasting with orhpaned image input

Change-Id: I65f19e462aa5aa9bebbdc3b748edb2d52e981d21