BUG:  Fixed colormap definition which was incorrect.
The jet colormapping function was incorrect in that the
shape of the red, green, and blue profiles were slightly
off.  The corrected version expands the center of the
spectrum.

This requires an update to
ITK/Testing/Data/Baseline/Review/RGBColormapTest_jet.png

To ensure that the jet colormap was consistent with the
colormap defined by matlab, I plotted the rgb profiles on the
same plot to ensure that they are as close to identical
as possible.

Request build: all

Change-Id: Ifcb4a82d1bdeecd09ca1df6480c8e21d98a3bfaa