STYLE:  and DOC:  corrected misspelling of Wiener, correct doc.
I had misspelled Wiener (my apologies to Prof. Wiener) and this
change replaces all instances of "Weiner" with "Wiener" in both
the N4 classes as well as the test .cxx file.  Also,
while I was at it, I corrected some of the incorrect documentation
which was showing up on Doxygen (a simple cut and paste had
replaced all instances of "N3" with "N4" and this change also
corrects that).

I compiled and built the changes on my machine (Mac OSX 10.6) and
ran the two N4 tests which passed.

Request build: all

Change-Id: I3433d204b00b8f8e87d67654d180fdb1a338ca18