BUG: Bad regression compare order pollutes Baseline directory.
If the test driver detects a baseline regression, it stores its
temporary test and difference image files in the directory of the
second file of the --compare argument. This should be the filename of
the generated test image. Some add_test commands had the order
reversed. If those tests failed the regression test, they wrote their
temp output files into Test/Data/Baselines/...

Change-Id: Ic8ae0cc5375b9751e16f01b11ab183bed9722c2c