BUG:  Remove unsupported cygwin compiler (11659)
We do not have any dashboards for cygwin. And it is not
being supported by the development team.

We recommend the use of the more compiant mingw for a gcc
compiler on windows.

This patch removes specific support for cygwin. Once
cygwin reaches its goal of being a posix system, then itk should work
with no if cygwin stuff at all. The next cygwin release of CMake will
treat the platform like posix.

Change-Id: I2cc6a21ed18f2652d580053b9fece64697818226