ENH: make possible to switch from system gccxml to EP in cmake gui
By default, GCCXML is set to the path of the gccxml built with an
external project. If the user choose to use the system gccxml, cmake
doesn't search for it because GCCXML is already set.
To avoid that, we delete GCCXML if it is equal to the gccxml path in
the external project before trying to find it with the usual
find_package() macro.

Change-Id: I53443b624ab7efa219f28bbacaae948a54dfdd43