ENH: Remove BasicFilters dependency from reg. test.
The current regression testing code uses filters that are in BasicFilters.
During the modularization, this dependency prevented the modules from
implementing regression tests that compare output images to baseline
images.

With the current modification, clones of the image comparison filters
have been created, that do not depend on the ImageToImage filter, and
provide equivalent functionalities. This refactoring introduces

itkTestingComparisonImageFilter        equivalent to itkDifferenceImageFilter
itkTestingExtractSliceImageFilter      equivalent to itkExtractImageFilter
itkTestingStretchIntensityImageFilter  equivalent to itkRescaleIntensityImageFilter

The new classes introduce a certain level of code redundancy, that
unfortunately is needed in order to break the module dependencies.

These Filters are not itended to be used by ITK Application developers.
In order to discourage widespread use, we have put them in a itk::Testing
namespace, and added the "Testing" prefix to their file names.

Change-Id: I554ca77e6eac6b2634c920e68e055be426188c52