ENH: moved StreamingImageIOBase and VTKImageIO out of review
VTKImageIO2 replaces VTKImageIO this adds streaming capabilities and
support for tensors to the class. It's reusabe base class is also
moved out of review.

Currently there are no vtk image files in the data submodule, once
this is added, the TODOs in the Testing/IO/CMakeLists.txt will be
addressed.

Change-Id: Ib778e274fce50d9b739b1648c49f24bf91c4fef2