BUG: Fix ImageFunction::IsInsideBuffer(ContinuousIndexType&) for NaN.
ENH: Add test for ImageFunction (in Testing/Code/Review).

ImageFunction::IsInsideBuffer(ContinuousIndexType& index) was not
catching NaN's in the index. It now does, and returns false if present.

There was no test for ImageFunction so one was added, testing the
behavior for a range of inputs, including NaN and Inf.

Change-Id: I7e55809c2ac32d658772a01e7a5584301d828bbe