BUG: Missing assignment operator.
Most of the iterators have the assignment operator implemented, but ImageRandomNonRepeatingConstIteratorWithIndex was missing it, and it was required because the m_Permutation object is not POD.

A test was added to create the object and then use the assignment operator. The test previously produced a segfault on the call to GoToBegin().

Change-Id: I0b6c89f1f50c47a9a0a5e95477f5143e4b5bf1b1