ENH: add ConvertLabelMapFilter
This filters allows to convert a label map to a different kind of
label object.

It may also be interesting, if possible, to simply dynamically cast
the LabelObjects in the output label map. In that case, it would even
be possible to run the filter in place, at a very small cost.

Change-Id: I633245f552bfd42938c8c5c077efb3359a98a8ae