ENH: itkImageRegionTest - added thorough tests for IsInside, including NaN.
DOC: Added comment to ImageRegion::IsInside( ContinousIndex ) that it
does not properly handle NaN inputs. It returns the correct value, but
not for the right reason.

ImageRegionTest was not testing the IsInside method expect for testing
regions. Added test for integer and continous index types, including
testing that NaN and infinity are handled properly.

Change-Id: I0aebe029eff9218c415cc95ce840ee96bcc4568a