ENH: Remove stray modularization manifest entries
The ITKConfig.cmake.in, UseITK.cmake.in, and itkGenerateITKConfig.cmake
files are part of the old monolithic build system.  They will need to be
re-written after modularization to perform the equivalent function.

The Utilities/FindLIBXML2.cmake file has been removed from the tree.

The Utilities/itkExtHdrs/fdstream.hpp certainly does not belong in the
non-existent Hooks module.  We will need to find a new home for it when
we refine the post-modularization build system.