BUG: baseline image needs to go before the test image with --compare
The first image after the --compare is the baseline image, the second
is the test image. The output off the diff and the other png images
are based on the path of the test image. If this is done in the wrong
order then the Testing data appears to be modified on the dashboard.

Change-Id: I4ae30849cd0c4c98d810c4c52d5739bbf1d0002d