ENH: Configure built-in IO Factories Registration.
After the modularization, the IO factories that used to be registered by
default, are no longer available. With this commit, an infrastructure is
introduced for assisting applications to automatically register the factories
that used to be registered by default in ITK 3.20 (minus the ImageIO formats
that were deprecated in ITKv4).

This patch still need to be customized to set the list of ImageIO factories
based on the collection of IO modules that were enabled in ITK at configuration
time.

Change-Id: I323bd1b9ee53c9811c6e04740d20e6e148a15e7d