ENH: Register IO factory if its module is ON.
In the previous commit, to ease the use of image file IO,  all builtin
IO factories are registered regardless whether the corresponding
IO modules are turned on of off.

In this patch, a certain IO factory get registered only if the corresponding
IO module is turned on during the configuration step, e.g. register PNG IO factory
with the image file reader/writer when Module_ITK-IO-PNG option is set to be ON.

In order to match the pre-modularization state of ITK, here we turn ON the IO
modules that used to be registered by default in ITK 3.20.

Change-Id: Id172cfba60383c26ef3c56be0bff0bc6953a5f54