BUG: Restore settings to CTestCustom
Accidentally changed behavior during the modularization effort.
This commit restores the pre-modularization settings, and also
updates some of the coverage exclusion paths to account for the
new names in the modularized source tree.

This commit also eliminates one of the two CTestCustom source
files in favor of the preferred one according to this blog post:
http://www.kitware.com/blog/home/post/27

Change-Id: I3a7b0c71779974785c0f5287a880b5d182e1ac18