ENH: Restore check for Testing/Data
Restore the ITK_DATA_ROOT setup code originally created by commit
4d785447 (Require Testing/Data for BUILD_(TESTING|EXAMPLES), 2011-01-31)
but removed during modularization.

If the submodule is not present provide instructions to get it.  Make
ITK_DATA_ROOT a local variable only.  It does not need to be a cache
variable because there is only one place the data may be located.