ENH: Remove unused TestNO_ICC_IDYNAMIC_NEEDED code
This code was used to conditionally add "-i_dynamic" to Intel compiler
builds.  We no longer support versions of that compiler old enough to
need the flag.  Use of this file was removed during modularization.

Use of -i_dynamic was originally added by commit bd224eaa (Work-around
for Intel compiler bug, 2003-12-22) and later made conditional by
commit 22e95976 (Enhanced testing for intel ICC compiler so that
-i_dynamic is only used when absolutly necessary, 2004-07-02) and
commit bfc1ee78 (Fixed cmake testing for intel compiler -i_dynamic
flag, 2004-07-06) and commit 57173789 (The logic for when to include
the -i_dynamic flag was reversed, 2004-07-06).