ENH: Restore MINC2 option in ITK-Review module
MINC2 support was originally added by commit 1d0dbd07 (Adding MINC2 to
the Review directory, 2007-08-27) but the option ITK_USE_MINC2 was
dropped during modularization.  Remove the option from itkConfigure.h as
it is not needed in C++ code.  Restore the option and the supporting
code in the ITK-Review module where it is needed.