BUG: FileFreeTestLib config
After modularization, the location of the FileFreeTestLib was not set
properly. This patch sets the LIBRARY_OUTPUT_DIRECTORY and also uses
the MODULE property for the add_library command. From the cmake
documentation, "MODULE libraries are plugins that are not linked into
other targets but may be loaded dynamically at runtime using
dlopen-like functionality."

COMP: itkInternationalizationIOHelpers comppile errors

Apparently, this code was never compiled. This patch fixes the compile
errors and corrects spelling errors in some method names.

Change-Id: I0d9111bb0edafd9a787bcbacc3255ae22f9f7d16