COMP: Shared lib suffix on Mac is .so or .dylib.
Shared libraries on the Mac may have either a .so or .dylib suffix.

Quoting Brad King,
"This is the difference in default name between using add_library with
SHARED v. MODULE.  At one time on the mac there was a distinction between
linkable libraries (.dylib) and loadable modules (.so) that could be
loaded with the predecessor to dlopen.  That is the original reason that
CMake makes a distinction between SHARED and MODULE libraries.  Now we
use it to distinguish linkable shared libraries from plugin modules on
all platforms."

Change-Id: I585471da148409a644178d3216fdef37ef660737