DOC: Use default doxygen style.
Somehow the css is not used (or detected?) when CREATE_SUBDIRS is on.
So let's use the standard Doxygen css, and default HTML_HEADER and
HTML_FOOTER.

Remove corresponding footer and css.

Make the documentation look similar to the VTK one's.

Remove the graphical class hierarchy, keep the textual one.

Modify header to make it more modern, but also to keep javascript for
online documentation editing.

Change-Id: Ia356b44d4ff0cc65d08ed21a5b694341faaa99ea