ENH: Avoid image copying in ImageSeriesReader
BUG: ImageSeries did not support VectorImages this is now fixed ( and
verified in SimpleITK)

The internal ImageFileReader can be told to load data into a section
of the output image already allocated by the ImageSeriesReader.
This reduces the time that it takes to load a series.

There are a great number of special cases which need to be taken into
consideration with this code, including streaming, reading a single
image, and reading an image which is a greater dimension then the
output etc. This filter may have too many features to manage.

This issue was pointed out in the developers mailing list by
Roger Bramon Feixas.

Change-Id: I63ccc59c29deb7a9d51ac905ddc8cfb43a9d0ad1