ENH: Register a group of IO factories for apps.
Currently, for an outside applcation(e.g. SimpleITK) to use several
image formats, users need to manually turn on the CMake variable
"Module_IO-<xxx>" for each image format when buiding ITK.
If a user turns on the entire IO Group via CMake var "ITKGroup_IO",
the IO modules are included in the built tree, but the IO factories
are not registered, which prevents the application to actually use
any image formats via ImageIOReader/Writer.

This commit solves this problem by correcting the conditions for registering
the IO factories in UseITK.cmake. When a user turns ON "ITKGroup_IO", all
image format modules are enabled and their corresponding "ITK-IO-<xxx>_LOADED"
are turned ON (in ITKConfigure.cmake.in).  "Module_ITK-IO-<xxx>" only reflects
users' input and remain unchanged. Therefore, "Module_ITK-IO-<xxx>" is no
longer needed in ITKConfig.cmake.in, and "ITK-IO-<xxx>_LOADED" is used to
register the IO factoires instead.

Change-Id: I65794a3610b9fbc67cdfb7ebd902af24d202f9b2