DOC: Add .mailmap to fix spelling of Gaëtan Lehmann
There are a few commits in history that use author name "Gaetan Lehmann"
created by a typo in the table entry the ITK Topic Stage uses to credit
merge commit authorship.  The typo has been corrected but the old
commits cannot be rewritten.  Add a .mailmap entry to map this author
name to the preferred spelling "Gaëtan Lehmann".  This affects the
output of "git shortlog" and "git blame".

Change-Id: I918e515012bf67c71b609ad19217e35ec38eb769