BUG: Fixes to Test Driver.
Fixed invocation of processes when not needed.

Recent changes to the itkTestDriverInclude.h that were introduced
to support the wrapping testing by making possible to call from
the test driver an external executable (e.g. Python interpreter)
resulted in unnecessary attempts to interpret every set of test
arguments as a potential external executable.

In this patch, we introduce an explicit flag "--process" to be
used to indicate when it is the developer's intention to actually
call an external program.

The ImageIOFactory::RegisterBuiltInFactories() was been removed
in a previous commit. Here we remove the declaration that was
left behind. Otherwise wrapping tests will segfault at runtime
due to the absence of this symbol in the shared libraries.

Change-Id: Ia8bd13eb88b8ef361e6e3d99a8a943e38002d3f1