ENH: Hide ITK module options when they are redundant
Each Module_* cache option exists to allow the user to request that its
specific module be built.  It does not indicate whether the module will
be built as a dependency of another module.  A module may build even if
its cache option is off.

We cannot force the options on or off based on the dependencies because
doing so erases information about the user's intent.  For example,
assume the user switches on a module that depends on several other
modules and we force the options for them to be on.  Now if the user
switches off the original module we have no idea whether its
dependencies were originally on or off.  The user would then have to
manually switch all of them back to the intended value.

The configuration output already specifies exactly what modules are
going to be built and why they are on (all modules are on, module is a
member of an enabled group, explicit request, or dependency of others).
Therefore we do not need to present this information in the CMake cache.

Note that if a module is on for any reason besides an explicit request
from its option then the option for the module itself is redundant.
Hide the option for the module in this case.  Furthermore, if
ITK_BUILD_ALL_MODULES is ON, hide the group options altogether.  This
approach keeps the set of cache options focused on decisions currently
available to the user.

Change-Id: Ie041f2c54a8ba7ac74b2932061c3228d63c9b343