ENH: Add HDF5 Image I/O
Use the HDF5 library to read & write images.

These are suitable for using with e.g. Matlab; the
pixel data is stored in an appropriately sized
N-d data array. If voxels have ComponentSize > 1,
they're stored as N+1-D array, and each voxel is contiguous
in the fastest moving dimension.

NOTE: The changes to NIFTI code are just changes needed so
that NIFTI and HDF5 can share some template code for reading
& writing images.

Change-Id: I06611be5dfa94467738cbac438af394f92758931

BUG: a;lsdfkj

Change-Id: I6146f6ee2c3866d183b798146de203ee30c9b429